webhook/requirements.txt
# Fix #21: pin versions to prevent silent breakage on reinstall
fastapi==0.115.6
uvicorn[standard]==0.32.1
pyyaml==6.0.2
